+++ elmyra.ip.access.epo/elmyra/ip/access/epo/static/js/app/models/ops.js
@@ -542,6 +542,19 @@ OpsExchangeDocument = Backbone.Model.extend({
             return results;
         },
 
+        has_fulltext: function() {
+            /*
+            3.1.2. Fulltext inquiry and retrieval including description or claims
+                   Note, Currently full texts (description and/or claims) are only available for the following
+                   authorities: EP, WO, AT, CA (claims only), CH.
+
+               -- http://documents.epo.org/projects/babylon/eponet.nsf/0/7AF8F1D2B36F3056C1257C04002E0AD6/$File/OPS_RWS_ReferenceGuide_version1210_EN.pdf
+            */
+            var country = this['@country'];
+            var countries_allowed = ['EP', 'WO', 'AT', 'CA', 'CH'];
+            return _(countries_allowed).contains(country);
+        },
+
     },
 
     select: function() {